third_party/blink/web_tests/external/wpt/loading/lazyload/below-viewport-image-loading-lazy-load-event.tentative-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Below-viewport loading=lazy images do not block the window load event when scrolled into viewport
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
FAIL Below-viewport loading=lazy images do not block the window load event when scrolled into viewport assert_true: The window load event should fire before the below-viewport loading=lazy images load expected true got false
Harness: the test ran to completion.
